Rename base dir from deadtrees to src
