+++ conf/callbacks/wandb_callbacks.yaml
@@ -3,6 +3,6 @@ defaults:
 
 
 watch_model_with_wandb:
-    _target_: deadtrees.callbacks.wandb_callbacks.WatchModelWithWandb
+    _target_: src.callbacks.wandb_callbacks.WatchModelWithWandb
     log: "all"
     log_freq: 100
