--- deadtrees/network/__init__.py
@@ -1,5 +1,0 @@
-from deadtrees.network.segmodel import SemSegment  # noqa: F401
-
-__all__ = [
-    "SemSegment",
-]
