--- deadtrees/visualization/__init__.py
@@ -1,5 +1,0 @@
-from deadtrees.visualization.helper import show  # noqa: F401
-
-__all__ = [
-    "show",
-]
