+++ setup.py
@@ -1,10 +1,10 @@
 import sys
 from itertools import chain
 
-from deadtrees import logger
-from deadtrees.constants import HOME_HTML, MODEL_CHECKPOINT_PATH, PACKAGE_DIR
-from deadtrees.version import __version__
 from setuptools import find_packages, setup
+from src import logger
+from src.constants import HOME_HTML, MODEL_CHECKPOINT_PATH, PACKAGE_DIR
+from src.version import __version__
 
 if not MODEL_CHECKPOINT_PATH.exists():
     # develop will be in argv if we do e.g. `pip install -e .`
@@ -52,7 +52,7 @@ setup(
     extras_require=EXTRAS,
     entry_points={
         "demo": [
-            "deadtrees=deadtrees.__main__:main",
+            "deadtrees=src.__main__:main",
         ],
     },
     package_data={
