+++ src/network/__init__.py
@@ -1,0 +1,5 @@
+from src.network.segmodel import SemSegment  # noqa: F401
+
+__all__ = [
+    "SemSegment",
+]
