+++ src/visualization/__init__.py
@@ -1,0 +1,5 @@
+from src.visualization.helper import show  # noqa: F401
+
+__all__ = [
+    "show",
+]
